+ Header fix and new text
Added sticky position to the header and make the scrollable area disappear under it.

Changed the text in the game page and made specific test if the user is in a small screen.

Still missing important update: scrollable main can never go over/under header, it should just disappear at a margin